feat: conditionally show the operations tile
