+++ client/app/data/dashboard/industry_user_approved_operator.json
@@ -1,10 +1,15 @@
 [
   {
     "title": "Select Operator",
-    "content": "View or update your operator information, which needs to be complete before applying for BORO ID's",
+    "content": "View or update your operator information",
     "links": [
       { "title": "Select Operator ➤", "href": "/dashboard/select-operator" }
     ]
+  },
+  {
+    "title": "My Operations",
+    "content": "Apply for a B.C. OBPS Regulated Operation ID",
+    "links": [{ "title": "My Operations ➤", "href": "/dashboard/operations" }]
   },
   {
     "title": "Report a Problem",
